+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/Post/dto/request/PostCreateRequest.java
@@ -1,0 +1,13 @@
+package com.example.SecondSeminar.Post.dto.request;
+
+import lombok.NonNull;
+
+public record PostCreateRequest(
+        @NonNull
+        String title,
+        @NonNull
+        String content,
+        @NonNull
+        Short categoryId
+) {
+}
