+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/Post/dto/request/PostUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.example.SecondSeminar.Post.dto.request;
+
+import lombok.NonNull;
+
+public record PostUpdateRequest(
+        @NonNull
+        String content
+) {
+}
