+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/Post/exception/PostException.java
@@ -1,0 +1,10 @@
+package com.example.SecondSeminar.Post.exception;
+
+import com.example.SecondSeminar.common.exception.BaseCustomException;
+import com.example.SecondSeminar.common.exception.ExceptionType;
+
+public class PostException extends BaseCustomException {
+    public PostException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
