+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/category/application/CategoryService.java
@@ -1,0 +1,21 @@
+package com.example.SecondSeminar.category.application;
+
+import com.example.SecondSeminar.category.domain.Category;
+import com.example.SecondSeminar.category.domain.CategoryId;
+import com.example.SecondSeminar.category.domain.CategoryJpaRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class CategoryService {
+
+    private final CategoryJpaRepository categoryJpaRepository;
+
+    public Category getByCategoryId(CategoryId categoryId) {
+        return categoryJpaRepository.findByIdOrThrowException(categoryId);
+    }
+
+}
