+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/category/domain/Category.java
@@ -1,0 +1,20 @@
+package com.example.SecondSeminar.category.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EmbeddedId;
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+public class Category {
+
+    @EmbeddedId
+    private CategoryId categoryId;
+
+    @Column(nullable = false)
+    private String name;
+}
