+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/category/exception/CategoryException.java
@@ -1,0 +1,10 @@
+package com.example.SecondSeminar.category.exception;
+
+import com.example.SecondSeminar.common.exception.BaseCustomException;
+import com.example.SecondSeminar.common.exception.ExceptionType;
+
+public class CategoryException extends BaseCustomException {
+    public CategoryException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
