+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/category/exception/CategoryExceptionType.java
@@ -1,0 +1,23 @@
+package com.example.SecondSeminar.category.exception;
+
+import com.example.SecondSeminar.common.exception.ExceptionType;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+public enum CategoryExceptionType implements ExceptionType {
+    NOT_FOUND_CATEGORY(HttpStatus.NOT_FOUND, "존재하지 않는 카테고리입니다.");
+
+    private final HttpStatus httpStatus;
+    private final String message;
+
+    @Override
+    public HttpStatus httpStatus() {
+        return httpStatus;
+    }
+
+    @Override
+    public String message() {
+        return message;
+    }
+}
