+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/common/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package com.example.SecondSeminar.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
