+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/common/exception/BaseCustomException.java
@@ -1,0 +1,16 @@
+package com.example.SecondSeminar.common.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class BaseCustomException extends RuntimeException {
+
+    private final ExceptionType exceptionType;
+
+    @Override
+    public String getMessage() {
+        return exceptionType.message();
+    }
+}
