+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/common/exception/ExceptionResponse.java
@@ -1,0 +1,15 @@
+package com.example.SecondSeminar.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class ExceptionResponse {
+    private String name;
+    private String message;
+
+    public static ExceptionResponse of(String name, String message) {
+        return new ExceptionResponse(name, message);
+    }
+}
