+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/common/exception/ExceptionType.java
@@ -1,0 +1,12 @@
+package com.example.SecondSeminar.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionType {
+    String name();
+
+    HttpStatus httpStatus();
+
+    String message();
+
+}
