--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.example.SecondSeminar.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
