--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/controller/dto/request/member/MemberCreateRequest.java
@@ -1,11 +1,0 @@
-package com.example.SecondSeminar.controller.dto.request.member;
-
-import com.example.SecondSeminar.domain.SOPT;
-
-public record MemberCreateRequest(
-        String name,
-        String nickname,
-        int age,
-        SOPT sopt
-) {
-}
