--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/controller/dto/request/member/MemberProfileUpdateRequest.java
@@ -1,9 +1,0 @@
-package com.example.SecondSeminar.controller.dto.request.member;
-
-import com.example.SecondSeminar.domain.Part;
-
-public record MemberProfileUpdateRequest(
-        int generation,
-        Part part
-) {
-}
