--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/controller/dto/request/post/PostCreateRequest.java
@@ -1,13 +1,0 @@
-package com.example.SecondSeminar.controller.dto.request.post;
-
-import lombok.NonNull;
-
-public record PostCreateRequest(
-        @NonNull
-        String title,
-        @NonNull
-        String content,
-        @NonNull
-        Short categoryId
-) {
-}
