--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/controller/dto/request/post/PostUpdateRequest.java
@@ -1,9 +1,0 @@
-package com.example.SecondSeminar.controller.dto.request.post;
-
-import lombok.NonNull;
-
-public record PostUpdateRequest(
-        @NonNull
-        String content
-) {
-}
