--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/domain/Post.java
@@ -1,51 +1,0 @@
-package com.example.SecondSeminar.domain;
-
-import com.example.SecondSeminar.domain.baseEntity.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Getter
-public class Post extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(columnDefinition = "TEXT", nullable = false)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @Column(name = "category_id", nullable = false)
-    private CategoryId categoryId;
-
-    @Builder
-    public Post(String title, String content, Member member, CategoryId categoryId) {
-        this.title = title;
-        this.content = content;
-        this.member = member;
-        this.categoryId = categoryId;
-    }
-
-    public void editContent(String content) {
-        this.content = content;
-    }
-}
