--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/domain/SOPT.java
@@ -1,22 +1,0 @@
-package com.example.SecondSeminar.domain;
-
-import static jakarta.persistence.EnumType.STRING;
-
-import jakarta.persistence.Embeddable;
-import jakarta.persistence.Enumerated;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class SOPT {
-    private int generation;
-
-    @Enumerated(value = STRING)
-    private Part part;
-
-}
