--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/domain/repository/CategoryJpaRepository.java
@@ -1,9 +1,0 @@
-package com.example.SecondSeminar.domain.repository;
-
-import com.example.SecondSeminar.domain.Category;
-import com.example.SecondSeminar.domain.CategoryId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryJpaRepository extends JpaRepository<Category, CategoryId> {
-
-}
