--- Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.SecondSeminar.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
