+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/member/domain/Part.java
@@ -1,0 +1,17 @@
+package com.example.SecondSeminar.member.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Part {
+    SERVER("서버"),
+    WEB("웹"),
+    ANDROID("안드로이드"),
+    IOS("iOS"),
+    PLAN("기획"),
+    DESIGN("디자인");
+
+    private final String name;
+}
