+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/member/dto/request/MemberCreateRequest.java
@@ -1,0 +1,11 @@
+package com.example.SecondSeminar.member.dto.request;
+
+import com.example.SecondSeminar.member.domain.SOPT;
+
+public record MemberCreateRequest(
+        String name,
+        String nickname,
+        int age,
+        SOPT sopt
+) {
+}
