+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/member/dto/request/MemberProfileUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.example.SecondSeminar.member.dto.request;
+
+import com.example.SecondSeminar.member.domain.Part;
+
+public record MemberProfileUpdateRequest(
+        int generation,
+        Part part
+) {
+}
