+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/member/exception/MemberException.java
@@ -1,0 +1,10 @@
+package com.example.SecondSeminar.member.exception;
+
+import com.example.SecondSeminar.common.exception.BaseCustomException;
+import com.example.SecondSeminar.common.exception.ExceptionType;
+
+public class MemberException extends BaseCustomException {
+    public MemberException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
