+++ Second Seminar/SecondSeminar/src/main/java/com/example/SecondSeminar/member/exception/MemberExceptionType.java
@@ -1,0 +1,23 @@
+package com.example.SecondSeminar.member.exception;
+
+import com.example.SecondSeminar.common.exception.ExceptionType;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+public enum MemberExceptionType implements ExceptionType {
+    NOT_FOUND_MEMBER(HttpStatus.NOT_FOUND, "존재하지 않는 멤버입니다.");
+
+    private final HttpStatus httpStatus;
+    private final String message;
+
+    @Override
+    public HttpStatus httpStatus() {
+        return httpStatus;
+    }
+
+    @Override
+    public String message() {
+        return message;
+    }
+}
